Organized the `library` module for future changes following clean architecture project structure
