+++ app/src/main/java/com/yausername/youtubedl_android_example/CommandExampleActivity.java
@@ -17,7 +17,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
 
 import com.yausername.youtubedl_android.YoutubeDL;
-import com.yausername.youtubedl_android.YoutubeDLRequest;
+import com.yausername.youtubedl_android.data.local.YoutubeDLRequest;
 
 import java.util.Collections;
 import java.util.regex.Matcher;
