+++ app/src/main/java/com/yausername/youtubedl_android_example/StreamingExampleActivity.java
@@ -15,7 +15,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.devbrackets.android.exomedia.listener.OnPreparedListener;
 import com.devbrackets.android.exomedia.ui.widget.VideoView;
 import com.yausername.youtubedl_android.YoutubeDL;
-import com.yausername.youtubedl_android.YoutubeDLRequest;
+import com.yausername.youtubedl_android.data.local.YoutubeDLRequest;
 
 import io.reactivex.Observable;
 import io.reactivex.android.schedulers.AndroidSchedulers;
