+++ library/src/main/java/com/yausername/youtubedl_android/domain/UpdateStatus.kt
@@ -1,0 +1,5 @@
+package com.yausername.youtubedl_android.domain
+
+enum class UpdateStatus {
+    DONE, ALREADY_UP_TO_DATE
+}
