+++ library/src/main/java/com/yausername/youtubedl_android/domain/model/videos/VideoThumbnail.kt
@@ -1,4 +1,4 @@
-package com.yausername.youtubedl_android.mapper
+package com.yausername.youtubedl_android.domain.model.videos
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
